backend/pangaea/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-appconf==1.0.5
django-compressor==4.1
django-sass-processor==1.2.1
djangorestframework==3.13.1
idna==3.3
libsass==0.21.0
numpy==1.23.3
pandas==1.4.4
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.2.1
rcssmin==1.1.0
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
